Fixed compiler warnings

+++ src/buffer_test.cc
@@ -136,7 +136,7 @@ TEST_F(BufferTest, GetHistogramForChannelGradient) {
   // Creates 10 RBGA pixel values with the blue channels ranging from 0 to 255.
   // Every value gets multiplied by 25 to create a gradient
   std::vector<Value> values(40);
-  for (uint64_t i = 0; i < values.size(); i += 4)
+  for (uint32_t i = 0; i < values.size(); i += 4)
     values[i + 2].SetIntValue(i / 4 * 25);
 
   Buffer b(BufferType::kColor);
@@ -144,7 +144,7 @@ TEST_F(BufferTest, GetHistogramForChannelGradient) {
   b.SetData(values);
 
   std::vector<uint64_t> bins = b.GetHistogramForChannel(2, 256);
-  for (uint64_t i = 0; i < values.size(); i += 4)
+  for (uint32_t i = 0; i < values.size(); i += 4)
     EXPECT_EQ(1, bins[i / 4 * 25]);
 }
 
@@ -156,7 +156,7 @@ TEST_F(BufferTest, GetHistogramForChannelAllBlack) {
   Format fmt(type.get());
 
   std::vector<Value> values(40);
-  for (uint64_t i = 0; i < values.size(); i++)
+  for (uint32_t i = 0; i < values.size(); i++)
     values[i].SetIntValue(0);
 
   Buffer b(BufferType::kColor);
@@ -165,7 +165,7 @@ TEST_F(BufferTest, GetHistogramForChannelAllBlack) {
 
   for (uint8_t i = 0; i < 4; i++) {
     std::vector<uint64_t> bins = b.GetHistogramForChannel(i, 256);
-    for (uint64_t y = 0; y < values.size(); y++)
+    for (uint32_t y = 0; y < values.size(); y++)
       EXPECT_EQ(10, bins[0]);
   }
 }
@@ -178,7 +178,7 @@ TEST_F(BufferTest, GetHistogramForChannelAllWhite) {
   Format fmt(type.get());
 
   std::vector<Value> values(40);
-  for (uint64_t i = 0; i < values.size(); i++)
+  for (uint32_t i = 0; i < values.size(); i++)
     values[i].SetIntValue(std::numeric_limits<uint8_t>::max());
 
   Buffer b(BufferType::kColor);
@@ -187,7 +187,7 @@ TEST_F(BufferTest, GetHistogramForChannelAllWhite) {
 
   for (uint8_t i = 0; i < 4; i++) {
     std::vector<uint64_t> bins = b.GetHistogramForChannel(i, 256);
-    for (uint64_t y = 0; y < values.size(); y++)
+    for (uint32_t y = 0; y < values.size(); y++)
       EXPECT_EQ(10, bins[255]);
   }
 }
@@ -202,7 +202,7 @@ TEST_F(BufferTest, CompareHistogramEMDToleranceFalse) {
 
   // Every value gets multiplied by 25 to create a gradient
   std::vector<Value> values1(40);
-  for (uint64_t i = 0; i < values1.size(); i += 4)
+  for (uint32_t i = 0; i < values1.size(); i += 4)
     values1[i].SetIntValue(i / 4 * 25);
 
   std::vector<Value> values2 = values1;
@@ -229,7 +229,7 @@ TEST_F(BufferTest, CompareHistogramEMDToleranceTrue) {
 
   // Every value gets multiplied by 25 to create a gradient
   std::vector<Value> values1(40);
-  for (uint64_t i = 0; i < values1.size(); i += 4)
+  for (uint32_t i = 0; i < values1.size(); i += 4)
     values1[i].SetIntValue(i / 4 * 25);
 
   std::vector<Value> values2 = values1;
@@ -254,7 +254,7 @@ TEST_F(BufferTest, CompareHistogramEMDToleranceAllBlack) {
   Format fmt(type.get());
 
   std::vector<Value> values1(40);
-  for (uint64_t i = 0; i < values1.size(); i++)
+  for (uint32_t i = 0; i < values1.size(); i++)
     values1[i].SetIntValue(0);
 
   std::vector<Value> values2 = values1;
@@ -278,7 +278,7 @@ TEST_F(BufferTest, CompareHistogramEMDToleranceAllWhite) {
   Format fmt(type.get());
 
   std::vector<Value> values1(40);
-  for (uint64_t i = 0; i < values1.size(); i++)
+  for (uint32_t i = 0; i < values1.size(); i++)
     values1[i].SetIntValue(std::numeric_limits<uint8_t>().max());
 
   std::vector<Value> values2 = values1;
